requirements.txt
Authlib==1.3.0
celery==5.3.6
Django==4.2.10
django-celery-results==2.5.1
django-prometheus==2.3.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
tabulate==0.9.0
